Fix date format logic

+++ app/src/main/java/com/github/ayltai/newspaper/util/DateUtils.java
@@ -33,14 +33,14 @@ public final class DateUtils {
 
         if (diff < 2 * DateUtils.MINUTE_MILLIS) return context.getString(R.string.time_ago_minute);
 
-        if (diff < 50 * DateUtils.MINUTE_MILLIS) return String.format(context.getString(R.string.time_ago_minutes), String.valueOf(Math.round(diff / DateUtils.MINUTE_MILLIS)));
+        if (diff < 50 * DateUtils.MINUTE_MILLIS) return String.format(context.getString(R.string.time_ago_minutes), String.valueOf(Math.round(diff / (double)DateUtils.MINUTE_MILLIS)));
 
         if (diff < 90 * DateUtils.MINUTE_MILLIS) return context.getString(R.string.time_ago_hour);
 
-        if (diff < 24 * DateUtils.HOUR_MILLIS) return String.format(context.getString(R.string.time_ago_hours), String.valueOf(Math.round(diff / DateUtils.HOUR_MILLIS)));
+        if (diff < 24 * DateUtils.HOUR_MILLIS) return String.format(context.getString(R.string.time_ago_hours), String.valueOf(Math.round(diff / (double)DateUtils.HOUR_MILLIS)));
 
         if (diff < 48 * DateUtils.HOUR_MILLIS) return context.getString(R.string.time_ago_day);
 
-        return String.format(context.getString(R.string.time_ago_days), String.valueOf(Math.round(diff / DateUtils.DAY_MILLIS)));
+        return String.format(context.getString(R.string.time_ago_days), String.valueOf(Math.round(diff / (double)DateUtils.DAY_MILLIS)));
     }
 }
